moves the async method to template method
